Adding release procedures to huborg
In a community based project, there are many moving parts. I wanted to
be clear about the steps used in releasing Huborg. The grant/revoke of
gem authority is something that can get sticky if you were commit a
bumped version and run `rake release`. You'll have tagged a release,
pushed to github, but not rubygems.

~enhancement