Highlighting links to API documentation
Prior to this commit, there was no indication as to where to go for
online API documentation. This commit remedies that.

As a bonus, I've switched the Badge Fury image from png to SVG, which
improved the legibility of the rendered image. The png was blurry and
looked odd beside the other SVGs.

Also tweaked a logging message to provide parity between the starting
and finishing task statement.