Adding Huborg::Client#each_pull_request_with_repo
As part of the duties of Samvera Tech facilitation, we wanted to
report on all pull requests. These pull requests span all of the
repositories in our samvera and samvera-labs.

I figured I may as well update the Huborg gem to provide some
automation.

As part of the change, I generalized the #fetch_rel_for logic to
now clarify that fetching rels is not limited to organizations.